Enable filtering of signals
Vertices may now choose whether or not they are used as the source or
the sink for a signal.

Implementing `transmits_signal(signal_params, transmission_params)`
allows vertices to avoid being the source for a signal.  Implementing
`accepts_signal(signal_params, transmission_params)` allows signals to
avoid being the sink for a signal.