Modify KeyspacesRegion to defer keyspace needs
Modify KeyspacesRegion so that instead of storing keyspaces it stores
references to signals which will later be allocated keyspaces and
dictionaries of values to load into those keyspaces. This makes later
automatic generation of keyspaces possible (to reduce the number of
unique identifiers used).

In one notable case a duplicated test is removed.