Use graph colouring to assign Signal IDs
Uses graph colouring to assign identifiers to signals (and hence keys to
multicast packets). This reduces the number of identifiers that
are required (previously each net had a unique identifier) and
consequently the number of bits in the key that are required to
represent the identifier.

The results of this are that:
 * Fewer bits in the key are required to represent IDs
 * The number of bits required is less strongly tied to the number of
   Signals in the model -- allowing larger models to be simulated without
   running out of bits in the key (necessary for Spaun)
 * Routing tables might be slightly smaller

Since this is a first pass there is scope for improvement. In particular
it is currently required that all of the multicast routes which arrive
at the same Nengo object have different keys - in principle this
constraint can be relaxed such that the same keys may be used by any
nets arriving at the same object given that (a) the nets do not route
differently from the same SpiNNaker node and (b) they do not target
different ports or filters in the object. Adjusting this should allow
the size of the filter routing table to be reduced, resulting in a
further performance improvement.