Key allocation for ensembles
Allows the graph colouring key allocation to be used for signals
arriving at ensembles. This boils down to the constraints on signals
being the union of the constraints required by the filter routing
regions AND the constraints required to ensure that signals which target
different sets of routing regions do not share a routing key.

In addition a small fix to the reporting of the routing key identifiers
and a fix to ensure that signals which have already been assigned
keyspaces are not included in the set of constraints reported by the
filter routing regions.