Minimise filter routing regions
Reduce the time taken to process some received multicast packets by
reducing the number of entries stored in the filter routing regions. The
"Critical Column First" minimisation method is used to achieve this as
it can, unlike "Ordered Covering" (see Rig), be used to minimise
unordered sets of routing keys.

Should further minimisation be necessary then a change to the way the
routing information is stored by each core may be necessary.