Rewrite filter routing to use pointers not indices
Currently the filter routing uses an index to look up the filter in
which to include the contribution of a routed packet. This look-up
occurs every time that the FIQ is called.

In this commit the routing structure is changed to point at the selected
filter, obviating the need to perform the look-up in every FIQ.