Assign cluster IDs using graph colouring
Cluster IDs are used to ensure the multicast nets representing signals
which originate from multiple chips can be differentiated if they take
different routes at the same router. Previously each chip was assigned a
unique identifier but this is unnecessary and wastes both routing table
entries and keyspace. This commit uses the graph colouring to try to
reduce the number of different cluster IDs applied, with the result the
fewer bits in the key and fewer routing tables are used to simulate the
same model.